Move yolox to object_detection module and update references and docs
