--- docs/docs/api-reference/models-megadetector_lite_yolox.md
@@ -1,3 +1,0 @@
-# zamba.models.megadetector_lite_yolox
-
-::: zamba.models.megadetector_lite_yolox
