--- docs/docs/api-reference/models-yolox_models.md
@@ -1,3 +1,0 @@
-# zamba.models.yolox_models
-
-::: zamba.models.yolox_models
