+++ docs/docs/api-reference/object_detection-yolox.md
@@ -1,0 +1,3 @@
+# zamba.object_detection.yolox
+
+::: zamba.object_detection.yolox
