+++ docs/mkdocs.yml
@@ -46,7 +46,8 @@ nav:
           - zamba.models.model_manager: "api-reference/models-model_manager.md"
           - zamba.models.slowfast_models: "api-reference/models-slowfast_models.md"
           - zamba.models.utils: "api-reference/models-utils.md"
-          - zamba.models.yolox_models: "api-reference/models-yolox_models.md"
+      - zamba.object_detection:
+          - zamba.object_detection.yolox: "api-reference/object_detection-yolox.md"
       - zamba.pytorch:
           - zamba.pytorch.dataloaders: "api-reference/pytorch-dataloaders.md"
           - zamba.pytorch.finetuning: "api-reference/pytorch-finetuning.md"
