+++ zamba/models/__init__.py
@@ -1,0 +1,2 @@
+from zamba.models.efficientnet_models import TimeDistributedEfficientNet  # noqa: F401
+from zamba.models.slowfast_models import SlowFast  # noqa: F401
