--- zamba/models/yolox_models/__init__.py
@@ -1,4 +1,0 @@
-from zamba.models.yolox_models.yolox_base import YoloXBase
-from zamba.models.yolox_models.yolox_nano import YoloXNano
-
-__all__ = ["YoloXBase", "YoloXNano"]
