+++ zamba/object_detection/__init__.py
@@ -1,0 +1,4 @@
+from zamba.object_detection.yolox.yolox_base import YoloXBase
+from zamba.object_detection.yolox.yolox_nano import YoloXNano
+
+__all__ = ["YoloXBase", "YoloXNano"]
