+++ zamba/pytorch_lightning/utils.py
@@ -19,21 +19,6 @@ from zamba.pytorch.dataloaders import get_datasets
 from zamba.pytorch.transforms import ConvertTHWCtoCTHW
 
 
-available_models = {}
-
-
-def register_model(cls):
-    """Used to decorate subclasses of ZambaVideoClassificationLightningModule so that they are
-    included in available_models."""
-    if not issubclass(cls, ZambaVideoClassificationLightningModule):
-        raise TypeError(
-            "Cannot register object that is not a subclass of "
-            "ZambaVideoClassificationLightningModule."
-        )
-    available_models[cls.__name__] = cls
-    return cls
-
-
 default_transform = transforms.Compose(
     [
         ConvertTHWCtoCTHW(),
